requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-mistralai
mistralai
chromadb
pypdf
python-dotenv
protobuf
langchain-text-splitters